contrib(profiling): move aop trace functionality into its own aop.trace package from aop
