+++ contribs/profiling/src/main/java/org/matsim/contrib/profiling/aop/trace/TraceProfilingAspect.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.profiling.aop;
+package org.matsim.contrib.profiling.aop.trace;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
